Migrate deprecated hash functions in the examples from pure .h files with static inline functions to proper .c and .h files.
For some reasons, gcc with -O3 does not handle very well all our static inline
functions in some situations, leading to incorrect results without any obvious
related undefined behaviour (as per C standard definition): this is specific
to gcc and clag handles this quite well.

In any case, migrating to .c and .h is cleaner.
