+++ src/examples/sig/rsa/Makefile
@@ -13,13 +13,19 @@ ifeq ($(WITH_DYNAMIC_LIBS),1)
 all:	rsa_dyn
 endif
 
-rsa:
-	$(CC) $(BIN_CFLAGS) -DRSA rsa.c ../../hash/hash.c $(ROOT_DIR)/src/external_deps/print.c $(ROOT_DIR)/src/external_deps/rand.c $(BIN_LDFLAGS) $(LIBSIGN) -o rsa
+libhash:
+	# Compile the hashes
+	cd ../../hash && EXTRA_CFLAGS="$(CFLAGS)" make libhash.a
+
+rsa: libhash
+	$(CC) $(BIN_CFLAGS) -DRSA rsa.c $(ROOT_DIR)/src/external_deps/print.c $(ROOT_DIR)/src/external_deps/rand.c $(BIN_LDFLAGS) ../../hash/libhash.a $(LIBSIGN) -o rsa
 
 # If the user asked for dynamic libraries, compile versions of our binaries against them
 ifeq ($(WITH_DYNAMIC_LIBS),1)
-rsa_dyn:
-	$(CC) $(BIN_CFLAGS) -DRSA rsa.c ../../hash/hash.c $(ROOT_DIR)/src/external_deps/print.c $(ROOT_DIR)/src/external_deps/rand.c $(BIN_LDFLAGS) $(LIBEC) -L$(BUILD_DIR) -lsign -o rsa_dyn
+rsa_dyn: libhash
+	# Compile the hashes
+	cd ../../hash && EXTRA_CFLAGS="$(CFLAGS)" make _li
+	$(CC) $(BIN_CFLAGS) -DRSA rsa.c $(ROOT_DIR)/src/external_deps/print.c $(ROOT_DIR)/src/external_deps/rand.c $(BIN_LDFLAGS) $(LIBEC) -L$(BUILD_DIR) ../../hash/libhash.a -lsign -o rsa_dyn
 endif
